+++ runtime/src/lib.rs
@@ -15,6 +15,7 @@ pub mod bank_utils;
 pub mod blockhash_queue;
 pub mod bloom;
 pub mod bucket_map_holder;
+pub mod bucket_map_holder_stats;
 pub mod builtins;
 pub mod commitment;
 pub mod contains;
